pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "UAVBiophysicalExtraction"
version = "0.1.0"
description = "Herramientas para la estimación biofísica usando UAVs"
readme = "README.md"
authors = [
    { name = "C.Minuesa", email = "cesar.minuesa@irta.cat" }
]
license = { file = "LICENSE" }
requires-python = ">=3.8"

dependencies = [
    "affine",
    "attrs",
    "Bottleneck",
    "branca",
    "Brotli",
    "certifi",
    "charset-normalizer",
    "click",
    "click-plugins",
    "cligj==0.7.2",
    "colorama",
    "contourpy",
    "cycler",
    "fiona==1.9.5",
    "folium",
    "fonttools",
    "GDAL",
    "geopandas",
    "idna",
    "imagecodecs",
    "imageio",
    "Jinja2",
    "joblib",
    "kiwisolver",
    "lazy_loader",
    "mapclassify",
    "MarkupSafe",
    "matplotlib==3.9.2",
    "mkl-service==2.4.0",
    "mkl_fft",
    "mkl_random",
    "networkx",
    "numexpr",
    "numpy==1.26.4",
    "packaging",
    "pandas==2.2.2",
    "pillow",
    "pyparsing",
    "pyproj",
    "PySocks",
    "python-dateutil",
    "pytz",
    "rasterio==1.3.10",
    "rasterstats==0.14.0",
    "requests",
    "Rtree",
    "scikit-image",
    "scikit-learn",
    "scipy==1.14.1",
    "setuptools==75.1.0",
    "shapely",
    "simplejson",
    "six",
    "snuggs",
    "threadpoolctl",
    "tifffile",
    "tzdata",
    "unicodedata2",
    "urllib3",
    "wheel==0.44.0",
    "win-inet-pton",
    "xyzservices"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
